pyproject.toml
[project]
name = "flask-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]
